requirements.txt
lxml==4.6.4
numpy==1.21.4
pandas==1.3.5
bs4==0.0.1
Pillow==8.4.0
python-Levenshtein==0.12.2
tqdm==4.62.3
click==8.0.3